refactor: rename from tosshin_sim into tosshin_gazebo_plugins
