+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.5)
 
-project(tosshin_sim)
+project(tosshin_gazebo_plugins)
 
 if(NOT CMAKE_C_STANDARD)
   set(CMAKE_C_STANDARD 99)
@@ -21,7 +21,7 @@ find_package(rclcpp REQUIRED)
 find_package(tosshin_interfaces REQUIRED)
 
 add_library(navigation_plugin SHARED
-  "src/tosshin_sim/navigation_plugin.cpp")
+  "src/navigation_plugin.cpp")
 
 target_include_directories(navigation_plugin PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -34,7 +34,7 @@ ament_target_dependencies(navigation_plugin
   tosshin_interfaces)
 
 install(TARGETS navigation_plugin
-  DESTINATION "share/${PROJECT_NAME}/plugins/${PROJECT_NAME}")
+  DESTINATION "share/${PROJECT_NAME}/plugins/tosshin")
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
