+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>tosshin_sim</name>
+  <name>tosshin_gazebo_plugins</name>
   <version>0.1.0</version>
   <description>Gazebo simulation version of Tosshin mobile robot navigation package</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
